Share code between DataHandler and Table components
Since the DataHandler was a modified copy-paste of the Table component,
a lot of code was common between them. A new TableDataHandler component
has been created to implement the common parts.
