--- web/html/src/components/data-handler.js
@@ -1,202 +1,0 @@
-/* eslint-disable */
-"use strict";
-
-const PropTypes = require('prop-types');
-const React = require("react");
-const createReactClass = require('create-react-class');
-const {PaginationBlock, ItemsPerPageSelector} = require('./pagination');
-const {SearchPanel} = require('components/table/SearchPanel');
-
-const DataHandler = createReactClass({
-  displayName: 'DataHandler',
-
-  propTypes: {
-    data: PropTypes.arrayOf(PropTypes.any).isRequired, // any type of data in and array, where each element is an item data
-    identifier: PropTypes.func.isRequired, // the unique key of the item
-    cssClassFunction: PropTypes.func, // a function that return a css class for each item
-    searchField: PropTypes.node, // the React Object that contains the filter search field
-    additionalFilters: PropTypes.node, // other filters to render but not handled here
-    initialItemsPerPage: PropTypes.number, // the initial number of how many item-per-page to show
-    selectable: PropTypes.bool, // enables item selection
-    onSelect: PropTypes.func, // the handler to call when the table selection is updated. if this function is not provided, the select boxes won't be rendered
-    selectedItems: PropTypes.array, // the identifiers for selected items
-    emptyText: PropTypes.string, // The message which is shown when there are no items to display
-    loading: PropTypes.bool, // if data is loading
-    loadingText: PropTypes.string, // The message which is shown when the data is loading
-  },
-
-  defaultEmptyText: t('There are no entries to show.'),
-  defaultLoadingText: t('Loading..'),
-
-  getInitialState: function() {
-    return {
-      currentPage: 1,
-      itemsPerPage: this.props.initialItemsPerPage || 15,
-      criteria: null,
-      selectedItems: this.props.selectedItems || [],
-      selectable: this.props.selectable,
-      loading: this.props.loading || false,
-    };
-  },
-
-  UNSAFE_componentWillReceiveProps: function(nextProps) {
-    this.onPageCountChange(nextProps.data, this.state.criteria, this.state.itemsPerPage);
-    this.setState({
-        selectedItems: nextProps.selectedItems || [],
-        selectable: Boolean(nextProps.selectable),
-        loading: Boolean(nextProps.loading) || false,
-    });
-  },
-
-  getLastPage: function(data, criteria, itemsPerPage) {
-    const itemCount = data.filter(this.getFilter(criteria)).length;
-
-    const lastPage = Math.ceil(itemCount / itemsPerPage);
-    return lastPage > 0 ? lastPage : 1;
-  },
-
-  getFilter: function(criteria) {
-    const searchField = this.props.searchField;
-    if (searchField) {
-      const filter = searchField.props.filter;
-      if (filter) {
-        return ((datum) => filter(datum, criteria));
-      }
-    }
-    return (datum) => true;
-  },
-
-  getProcessedData: function() {
-    return this.props.data.filter(this.getFilter(this.state.criteria));
-  },
-
-  onSearch: function(criteria) {
-    this.setState({criteria: criteria});
-    this.onPageCountChange(this.props.data, criteria, this.state.itemsPerPage);
-  },
-
-  onItemsPerPageChange: function(itemsPerPage) {
-    this.setState({itemsPerPage: itemsPerPage});
-    this.onPageCountChange(this.props.data, this.state.criteria, itemsPerPage);
-  },
-
-  onPageCountChange: function(data, criteria, itemsPerPage) {
-    const lastPage = this.getLastPage(data, criteria, itemsPerPage);
-    if (this.state.currentPage > lastPage) {
-      this.setState({currentPage: lastPage});
-    }
-  },
-
-  onPageChange: function(page) {
-    this.setState({currentPage: page});
-  },
-
-  setSelection: function(selection) {
-    if(this.props.onSelect) {
-      this.props.onSelect(selection);
-    }
-  },
-
-  render: function() {
-    const filteredData = this.getProcessedData();
-
-    const itemsPerPage = this.state.itemsPerPage;
-    const currentPage = this.state.currentPage;
-    const firstItemIndex = (currentPage - 1) * itemsPerPage;
-
-    const itemCount = filteredData.length;
-    const fromItem = itemCount > 0 ? firstItemIndex + 1 : 0;
-    const toItem = firstItemIndex + itemsPerPage <= itemCount ? firstItemIndex + itemsPerPage : itemCount;
-    const currItems = filteredData.slice(firstItemIndex, firstItemIndex + itemsPerPage);
-    const isEmpty = itemCount === 0;
-
-    const handleSelect = (id, sel) => {
-        let arr = this.state.selectedItems;
-        if(sel) {
-            arr = arr.concat([id]);
-        } else {
-            arr = arr.filter(i => i !== id);
-        }
-        this.setSelection(arr);
-    };
-
-    const dataItems = React.Children.toArray(this.props.children)
-        .map(item => React.cloneElement(item, {data: currItems, criteria: this.state.criteria}));
-
-    const handleSearchPanelClear = () => {
-        this.setSelection([]);
-    }
-
-    const handleSearchPanelSelectAll = () => {
-        const selected = this.state.selectedItems;
-        this.setSelection(selected.concat(
-            filteredData.map(d => this.props.identifier(d))
-                .filter(id => !selected.includes(id))));
-    }
-
-    const emptyText = this.props.emptyText || this.defaultEmptyText;
-    const loadingText= this.props.loadingText || this.defaultLoadingText;
-
-    return (
-      <div className="spacewalk-list">
-        <div className="panel panel-default">
-          <div className="panel-heading">
-            <div className="spacewalk-list-head-addons">
-            <SearchPanel
-              fromItem={fromItem}
-              toItem={toItem}
-              itemCount={itemCount}
-              criteria={this.state.criteria}
-              onSearch={this.onSearch}
-              onClear={handleSearchPanelClear}
-              onSelectAll={handleSearchPanelSelectAll}
-              selectedCount={this.state.selectedItems.length}
-              selectable={this.state.selectable}
-            >{this.props.searchField}
-              {
-                this.props.additionalFilters.map((filter, i) => <span key={'additional-filter-' + i}>{filter}&nbsp;</span>)
-              }
-            </SearchPanel>
-              <div className="spacewalk-list-head-addons-extra table-items-per-page-wrapper">
-                <ItemsPerPageSelector key="itemsPerPageSelector"
-                  currentValue={this.state.itemsPerPage}
-                  onChange={this.onItemsPerPageChange}
-                /> {t("items per page")}
-              </div>
-            </div>
-          </div>
-          { this.state.loading ?
-            <div className="panel-body text-center">
-              <i className='fa fa-spinner fa-spin fa-1-5x'></i>
-              <h4>{loadingText}</h4>
-            </div>
-            :
-            (
-              isEmpty ?
-              <div className="panel-body">
-                <h4>{emptyText}</h4>
-              </div>
-              :
-              <div className="table-responsive">
-                {dataItems}
-              </div>
-            )
-          }
-          <div className="panel-footer">
-            <div className="spacewalk-list-bottom-addons">
-              <PaginationBlock key="paginationBlock"
-                currentPage={this.state.currentPage}
-                lastPage={this.getLastPage(this.props.data, this.state.criteria, this.state.itemsPerPage)}
-                onPageChange={this.onPageChange}
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-    );
-  },
-});
-
-module.exports = {
-    DataHandler : DataHandler,
-}
