+++ web/html/src/components/table/Table.js
@@ -1,12 +1,8 @@
 // @flow
 import * as React from 'react';
-import { StatePersistedContext } from '../util/StatePersistedContext';
-import { ItemsPerPageSelector } from '../pagination';
-import { PaginationBlock } from '../pagination';
-import { SearchPanel } from './SearchPanel';
+import { TableDataHandler } from './TableDataHandler';
 import { SearchField } from './SearchField';
 import { Column } from './Column';
-import { Header } from './Header';
 
 type TableProps = {
   /** any type of data in an array, where each element is a row data */
@@ -39,286 +35,60 @@ type TableProps = {
   children: React.Node,
 };
 
-type TableState = {
-  currentPage: number,
-  itemsPerPage: number,
-  criteria?: string,
-  sortColumnKey: string | null,
-  sortDirection: number,
-  selectedItems: Array<any>,
-  selectable: boolean,
-  loading: boolean,
-};
+export function Table(props: TableProps): React.Node {
+  const {...allProps} = props;
+  const columns = React.Children.toArray(props.children)
+    .filter((child) => child.type === Column || child.type.displayName === "Column")
+    .map((child) => React.cloneElement(child));
+  return (
+    <TableDataHandler
+      columns={columns}
+      {...allProps}
+    >
+    {
+      ({currItems, headers, handleSelect, selectable, selectedItems, criteria}) => {
+        const rows = currItems.map((datum, index) => {
+            const cells = React.Children.toArray(props.children)
+              .filter((child) => child.type === Column || child.type.displayName === "Column")
+              .map((column) => React.cloneElement(column, {data: datum, criteria: criteria})
+            );
 
-export class Table extends React.Component<TableProps, TableState> {
-  static defaultProps = {
-    selectable: false,
-  };
-
-  static contextType = StatePersistedContext;
-
-  constructor(props: TableProps) {
-    super(props);
-    this.state = {
-      currentPage: 1,
-      itemsPerPage: this.props.initialItemsPerPage || 15,
-      criteria: undefined,
-      sortColumnKey: this.props.initialSortColumnKey || null,
-      sortDirection: this.props.initialSortDirection || 1,
-      selectedItems: this.props.selectedItems || [],
-      selectable: this.props.selectable,
-      loading: this.props.loading || false,
-    };
-  };
-
-  UNSAFE_componentWillReceiveProps(nextProps: TableProps) {
-    this.onPageCountChange(nextProps.data, this.state.criteria, this.state.itemsPerPage);
-    this.setState({
-        selectedItems: nextProps.selectedItems || [],
-        selectable: Boolean(nextProps.selectable),
-        loading: Boolean(nextProps.loading) || false,
-    });
-  };
-
-  componentWillUnmount() {
-    if (this.context && this.context.saveState) {
-      this.context.saveState(this.state);
-    }
-  };
-
-  UNSAFE_componentWillMount() {
-    if (this.context && this.context.loadState) {
-      if (this.context.loadState()) {
-        this.setState(this.context.loadState());
-      }
-    }
-  };
-
-  getLastPage = (data: any, criteria?: string, itemsPerPage: number) : number => {
-    const rowCount = data.filter(this.getFilter(criteria)).length;
-
-    const lastPage = Math.ceil(rowCount / itemsPerPage);
-    return lastPage > 0 ? lastPage : 1;
-  };
-
-  getFilter = (criteria?: string): (datum: any) => boolean => {
-    const searchField = this.props.searchField;
-    if (searchField) {
-      const filter = searchField.props.filter;
-      if (filter) {
-        return ((datum) => filter(datum, criteria));
-      }
-    }
-    return (datum) => true;
-  };
-
-  getProcessedData = (): Array<any> => {
-    const comparators = React.Children.toArray(this.props.children)
-      .filter((child) => child.type.displayName === Column.name)
-      .filter((column) => column.props.columnKey === this.state.sortColumnKey)
-      .map((column) => column.props.comparator);
-
-    const comparator = comparators.length > 0 ?
-      comparators[0] : ((a, b, columnKey, sortDirection) => 0);
-
-    return this.props.data
-        .filter(this.getFilter(this.state.criteria))
-        .sort((a, b) => comparator(a, b, this.state.sortColumnKey, this.state.sortDirection));
-  };
-
-  onSearch = (criteria?: string): void => {
-    this.setState({criteria: criteria});
-    this.onPageCountChange(this.props.data, criteria, this.state.itemsPerPage);
-  };
-
-  onItemsPerPageChange = (itemsPerPage: number): void => {
-    this.setState({itemsPerPage: itemsPerPage});
-    this.onPageCountChange(this.props.data, this.state.criteria, itemsPerPage);
-  };
-
-  onPageCountChange = (data: Array<any>, criteria?: string, itemsPerPage: number): void => {
-    const lastPage = this.getLastPage(data, criteria, itemsPerPage);
-    if (this.state.currentPage > lastPage) {
-      this.setState({currentPage: lastPage});
-    }
-  };
-
-  onPageChange = (page: number): void => {
-    this.setState({currentPage: page});
-  };
-
-  onSortChange = (sortColumnKey?: string, sortDirection: number): void => {
-    this.setState({
-      sortColumnKey: sortColumnKey,
-      sortDirection: sortDirection
-    });
-  };
-
-  setSelection = (selection: any): void => {
-    if (this.props.onSelect) {
-      this.props.onSelect(selection);
-    }
-  };
-
-  render() {
-    const headers = React.Children.toArray(this.props.children)
-        .filter((child) => child.type.displayName === Column.name)
-        .map((column, index) => {
-            if (column.props.header) {
-                const sortDirection = column.props.columnKey === this.state.sortColumnKey ?
-                  this.state.sortDirection :
-                  0;
-                return <Header
-                    key={index}
-                    columnKey={column.props.columnKey}
-                    sortDirection={sortDirection}
-                    onSortChange={this.onSortChange}
-                    width={column.props.width}
-                    comparator={column.props.comparator}
-                    className={column.props.headerClass}>
-                        {column.props.header}
-                    </Header>;
-            } else {
-                return <Header key={index} className={column.props.headerClass}/>;
+            if (selectable) {
+              const checkbox = <Column key="check" cell={
+                <input type="checkbox"
+                    checked={selectedItems.includes(props.identifier(datum))}
+                    onChange={(e) => handleSelect(props.identifier(datum), e.target.checked)}
+                />
+              }/>;
+              cells.unshift(checkbox);
             }
+
+            const rowClass = props.cssClassFunction ? props.cssClassFunction(datum, index) : "";
+            const evenOddClass = (index % 2) === 0 ? "list-row-even" : "list-row-odd";
+            return (
+              <tr className={rowClass + " " + evenOddClass}
+                key={props.identifier(datum)}
+              >
+                {cells}
+              </tr>
+            );
         });
 
-    const filteredData = this.getProcessedData();
-
-    const itemsPerPage = this.state.itemsPerPage;
-    const currentPage = this.state.currentPage;
-    const firstItemIndex = (currentPage - 1) * itemsPerPage;
-
-    const itemCount = filteredData.length;
-    const fromItem = itemCount > 0 ? firstItemIndex + 1 : 0;
-    const toItem = firstItemIndex + itemsPerPage <= itemCount ? firstItemIndex + itemsPerPage : itemCount;
-    const currItems = filteredData.slice(firstItemIndex, firstItemIndex + itemsPerPage);
-    const currIds = currItems.map(item => this.props.identifier(item));
-    const isEmpty = itemCount === 0;
-
-    const handleSelectAll = (sel) => {
-        let arr = this.state.selectedItems;
-        if (sel) {
-            arr = arr.concat(currIds.filter(id => !arr.includes(id)));
-        } else {
-            arr = arr.filter(id => !currIds.includes(id));
-        }
-        this.setSelection(arr);
-    };
-
-    if (this.state.selectable) {
-        const allSelected = currIds.length > 0 && currIds.every(id => this.state.selectedItems.includes(id));
-        const checkbox = <Header key="check"><input type="checkbox" checked={allSelected} onChange={(e) => handleSelectAll(e.target.checked)}/></Header>;
-        headers.unshift(checkbox);
-    }
-
-    const handleSelect = (id, sel) => {
-        let arr = this.state.selectedItems;
-        if (sel) {
-            arr = arr.concat([id]);
-        } else {
-            arr = arr.filter(i => i !== id);
-        }
-        this.setSelection(arr);
-    };
-
-    const rows = currItems.map((datum, index) => {
-        const cells = React.Children.toArray(this.props.children)
-          .filter((child) => child.type.displayName === Column.name)
-          .map((column) => React.cloneElement(column, {data: datum, criteria: this.state.criteria})
+        return (
+          <table className="table table-striped vertical-middle">
+            <thead>
+              <tr>{headers}</tr>
+            </thead>
+            <tbody>
+              {rows}
+            </tbody>
+          </table>
         );
-
-        if (this.state.selectable) {
-          const checkbox = <Column key="check" cell={
-            <input type="checkbox"
-                checked={this.state.selectedItems.includes(this.props.identifier(datum))}
-                onChange={(e) => handleSelect(this.props.identifier(datum), e.target.checked)}
-            />
-          }/>;
-          cells.unshift(checkbox);
-        }
-
-        let rowClass = this.props.cssClassFunction ? this.props.cssClassFunction(datum, index) : "";
-        let evenOddClass = (index % 2) === 0 ? "list-row-even" : "list-row-odd";
-        return <tr className={rowClass + " " + evenOddClass} key={this.props.identifier(datum)} >{cells}</tr>;
-    });
-
-    const handleSearchPanelClear = () => {
-        this.setSelection([]);
+      }
     }
-
-    const handleSearchPanelSelectAll = () => {
-        const selected = this.state.selectedItems;
-        this.setSelection(selected.concat(
-            filteredData.map(d => this.props.identifier(d))
-                .filter(id => !selected.includes(id))));
-    }
-
-    const emptyText = this.props.emptyText || t('There are no entries to show.');
-    const loadingText = this.props.loadingText || t('Loading...');
-
-    return (
-      <div className="spacewalk-list">
-        <div className="panel panel-default">
-          <div className="panel-heading">
-            <div className="spacewalk-list-head-addons">
-            <SearchPanel
-              fromItem={fromItem}
-              toItem={toItem}
-              itemCount={itemCount}
-              criteria={this.state.criteria}
-              onSearch={this.onSearch}
-              onClear={handleSearchPanelClear}
-              onSelectAll={handleSearchPanelSelectAll}
-              selectedCount={this.state.selectedItems.length}
-              selectable={this.state.selectable}
-            >{this.props.searchField}
-            </SearchPanel>
-              <div className="spacewalk-list-head-addons-extra table-items-per-page-wrapper">
-                <ItemsPerPageSelector key="itemsPerPageSelector"
-                  currentValue={this.state.itemsPerPage}
-                  onChange={this.onItemsPerPageChange}
-                /> {t("items per page")}
-              </div>
-            </div>
-          </div>
-          { this.state.loading ?
-            <div className="panel-body text-center">
-              <i className='fa fa-spinner fa-spin fa-1-5x'></i>
-              <h4>{loadingText}</h4>
-            </div>
-            :
-            (
-              isEmpty ?
-              <div className="panel-body">
-                <h4>{emptyText}</h4>
-              </div>
-              :
-              <div>
-                <div className="table-responsive">
-                  <table className="table table-striped vertical-middle">
-                    <thead>
-                      <tr>{headers}</tr>
-                    </thead>
-                    <tbody>
-                      {rows}
-                    </tbody>
-                  </table>
-                </div>
-              </div>
-            )
-          }
-          <div className="panel-footer">
-            <div className="spacewalk-list-bottom-addons">
-              <PaginationBlock key="paginationBlock"
-                currentPage={this.state.currentPage}
-                lastPage={this.getLastPage(this.props.data, this.state.criteria, this.state.itemsPerPage)}
-                onPageChange={this.onPageChange}
-              />
-            </div>
-          </div>
-        </div>
-      </div>
-    );
-  }
+    </TableDataHandler>
+  );
+};
+Table.defaultProps = {
+  selectable: false,
 };
